3-scaling/requirements.txt
azure-servicebus==7.11.4
azure-identity